restructure class as separate client/server caching
