+++ java/client/src/main/java/asyncCache/client/AsyncMemManager.java
@@ -1,4 +1,4 @@
-package asyncMemManager.common;
+package asyncCache.client;
 
 import java.time.LocalTime;
 import java.time.temporal.ChronoField;
@@ -12,18 +12,21 @@ import java.util.concurrent.PriorityBlockingQueue;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
-import asyncMemManager.common.AsyncMemManager.ManagedObjectBase.KeyLock;
-import asyncMemManager.common.AsyncMemManager.ManagedObjectBase.ManageKeyLock;
-import asyncMemManager.common.AsyncMemManager.ManagedObjectBase.ReadKeyLock;
+import asyncCache.client.AsyncMemManager.ManagedObjectBase.KeyLock;
+import asyncCache.client.AsyncMemManager.ManagedObjectBase.ManageKeyLock;
+import asyncCache.client.AsyncMemManager.ManagedObjectBase.ReadKeyLock;
+import asyncMemManager.common.Configuration;
+import asyncMemManager.common.ManagedObjectQueue;
 import asyncMemManager.common.di.BinarySerializer;
+import asyncMemManager.common.di.IndexableQueuedObject;
 
-public class AsyncMemManager implements asyncMemManager.common.di.AsyncMemManager, AutoCloseable {
+public class AsyncMemManager implements asyncCache.client.di.AsyncMemManager, AutoCloseable {
 	
 	private Configuration config;
 	private asyncMemManager.common.di.HotTimeCalculator coldTimeCalculator;
 	private asyncMemManager.common.di.Persistence persistence;
-	private BlockingQueue<ManagedObjectQueue> candlesPool;
-	private List<ManagedObjectQueue> candles;
+	private BlockingQueue<ManagedObjectQueue<ManagedObjectBase>> candlesPool;
+	private List<ManagedObjectQueue<ManagedObjectBase>> candles;
 	private ConcurrentHashMap<UUID, ManagedObjectBase> keyToObjectMap;
 	private AtomicLong usedSize;
 	private Comparator<ManagedObjectBase> coldCacheNodeComparator = (n1, n2) -> n2.hotTime.compareTo(n1.hotTime);
@@ -44,17 +47,17 @@ public class AsyncMemManager implements asyncMemManager.common.di.AsyncMemManage
 		this.config = config;
 		this.coldTimeCalculator = coldTimeCalculator;
 		this.persistence = persistence;
-		this.keyToObjectMap = new ConcurrentHashMap<>(this.config.initialSize);
-		this.candlesPool = new PriorityBlockingQueue<>(this.config.candlePoolSize, 
+		this.keyToObjectMap = new ConcurrentHashMap<>(this.config.getInitialSize());
+		this.candlesPool = new PriorityBlockingQueue<>(this.config.getCandlePoolSize(), 
 														(c1, c2) -> Integer.compare(c1.size(), c2.size()));
-		this.candles = new ArrayList<>(this.config.candlePoolSize);
+		this.candles = new ArrayList<>(this.config.getCandlePoolSize());
 		
-		int initcandleSize = this.config.initialSize / this.config.candlePoolSize;
-		initcandleSize = initcandleSize > 0 ? initcandleSize : this.config.initialSize;
+		int initcandleSize = this.config.getInitialSize() / this.config.getCandlePoolSize();
+		initcandleSize = initcandleSize > 0 ? initcandleSize : this.config.getInitialSize();
 		// init candle pool
-		for(int i = 0; i < config.candlePoolSize; i++)
+		for(int i = 0; i < config.getCandlePoolSize(); i++)
 		{
-			ManagedObjectQueue candle = new ManagedObjectQueue(initcandleSize, this.coldCacheNodeComparator); // thread-safe ensured by candlesPool
+			ManagedObjectQueue<ManagedObjectBase> candle = new ManagedObjectQueue<>(initcandleSize, this.coldCacheNodeComparator); // thread-safe ensured by candlesPool
 			this.candlesPool.add(candle);
 			this.candles.add(candle);
 		}
@@ -112,7 +115,7 @@ public class AsyncMemManager implements asyncMemManager.common.di.AsyncMemManage
 			this.keyToObjectMap.put(managedObj.key, managedObj);		
 			
 			// put node to candle
-			ManagedObjectQueue candle = null;
+			ManagedObjectQueue<ManagedObjectBase> candle = null;
 			try {
 				candle = this.candlesPool.take();
 			} catch (InterruptedException e) {
@@ -137,7 +140,7 @@ public class AsyncMemManager implements asyncMemManager.common.di.AsyncMemManage
 			if (managedObj.containerCandle == null)
 			{
 				// put node to candle
-				ManagedObjectQueue candle = null;
+				ManagedObjectQueue<ManagedObjectBase> candle = null;
 				try {
 					candle = this.candlesPool.take();
 				} catch (InterruptedException e) {
@@ -169,7 +172,7 @@ public class AsyncMemManager implements asyncMemManager.common.di.AsyncMemManage
 			}			
 			
 			KeyLock lock = managedObj.lockManage();
-			ManagedObjectQueue candle = managedObj.containerCandle;
+			ManagedObjectQueue<ManagedObjectBase> candle = managedObj.containerCandle;
 			managedObj.containerCandle.removeAt(managedObj.candleIndex);
 			managedObj.containerCandle = null;
 			this.usedSize.addAndGet(-managedObj.estimatedSize);
@@ -184,7 +187,7 @@ public class AsyncMemManager implements asyncMemManager.common.di.AsyncMemManage
 	
 	private boolean isOverCapability()
 	{
-		return this.usedSize.get() > this.config.capacity;
+		return this.usedSize.get() > this.config.getCapacity();
 	}
 	
 	private volatile boolean waitingForPersistence = false;
@@ -199,7 +202,7 @@ public class AsyncMemManager implements asyncMemManager.common.di.AsyncMemManage
 		while (this.isOverCapability())
 		{
 			ManagedObjectBase coldestNode = null;
-			for (ManagedObjectQueue candle : this.candles)
+			for (ManagedObjectQueue<ManagedObjectBase> candle : this.candles)
 			{
 				ManagedObjectBase node = candle.peek();
 				if (node != null)
@@ -213,7 +216,7 @@ public class AsyncMemManager implements asyncMemManager.common.di.AsyncMemManage
 			
 			if (coldestNode != null)
 			{			
-				ManagedObjectQueue coldestCandle = coldestNode.containerCandle;
+				ManagedObjectQueue<ManagedObjectBase> coldestCandle = coldestNode.containerCandle;
 				if (coldestCandle != null)
 				{
 					while(!this.candlesPool.remove(coldestCandle))
@@ -258,34 +261,8 @@ public class AsyncMemManager implements asyncMemManager.common.di.AsyncMemManage
 		
 		this.waitingForPersistence = false;
 	}
-
-	public static class FlowKeyConfiguration
-	{	
-	}
 	
-	public static class Configuration
-	{
-		int initialSize;
-		int capacity;
-		int cleanupInterval;
-		int candlePoolSize;
-		Map<String, FlowKeyConfiguration> flowKeyConfig = new HashMap<>();
-
-		public Configuration(int capacity, 
-								int initialSize, 
-								int cleanupInterval,
-								int candlePoolSize,
-								Map<String, FlowKeyConfiguration> flowKeyConfig) 
-		{
-			this.capacity = capacity;
-			this.candlePoolSize = candlePoolSize >= 0 ? candlePoolSize : Runtime.getRuntime().availableProcessors();
-			this.initialSize = initialSize > 0 ? initialSize : 100;
-			this.cleanupInterval = cleanupInterval;
-			this.flowKeyConfig = flowKeyConfig;
-		}
-	}
-	
-	abstract class ManagedObjectBase
+	abstract class ManagedObjectBase implements IndexableQueuedObject
 	{
 		/***
 		 * key value to lookup object, this is auto unique generated00
@@ -315,7 +292,7 @@ public class AsyncMemManager implements asyncMemManager.common.di.AsyncMemManage
 		/**
 		 * the candle contain this object, used for fast cleanup, removal
 		 */
-		ManagedObjectQueue containerCandle;
+		ManagedObjectQueue<ManagedObjectBase> containerCandle;
 		
 		/**
 		 * the index of object in candle, used for fast removal
@@ -362,6 +339,12 @@ public class AsyncMemManager implements asyncMemManager.common.di.AsyncMemManage
 		 * used for read/write locking this managed object. 
 		 */
 		private int accessCounter = 0;
+		
+		@Override
+		public void setIndexInQueue(int idx)
+		{
+			this.candleIndex = idx;
+		}
 		
 		/**
 		 * read locking, used for async flows access object, to ensure data not interfered
@@ -536,123 +519,5 @@ public class AsyncMemManager implements asyncMemManager.common.di.AsyncMemManage
 				this.managedObject.startTracking();
 			}
 		}
-	}
-	
-	/**
-	 * Custom Queue implementation for candle queue of managed objects.
-	 */
-	private static class ManagedObjectQueue{
-		private static final int MAX_ARRAY_SIZE = Integer.MAX_VALUE - 8;
-		
-		ManagedObjectBase[] queue;
-	    private int size = 0;
-	    private final Comparator<ManagedObjectBase> comparator;
-	    
-	    public ManagedObjectQueue(int initSize, Comparator<ManagedObjectBase> comparator) {
-	        this.queue = new ManagedObjectBase[initSize];
-	        this.comparator = comparator;
-	    }
-	    
-	    private void grow(int minCapacity) {
-	        int oldCapacity = queue.length;
-	        // Double size if small; else grow by 50%
-	        int newCapacity = oldCapacity
-	                + ((oldCapacity < 64) ? (oldCapacity + 2) : (oldCapacity >> 1));
-	        // overflow-conscious code
-	        if (newCapacity - MAX_ARRAY_SIZE > 0)
-	            newCapacity = hugeCapacity(minCapacity);
-	        queue = Arrays.copyOf(queue, newCapacity);
-	    }
-	    
-	    private static int hugeCapacity(int minCapacity) {
-	        if (minCapacity < 0) // overflow
-	            throw new OutOfMemoryError();
-	        return (minCapacity > MAX_ARRAY_SIZE) ? Integer.MAX_VALUE : MAX_ARRAY_SIZE;
-	    }
-	    
-	    public ManagedObjectBase peek() {
-	        return (size == 0) ? null : queue[0];
-	    }
-	    
-	    public ManagedObjectBase poll() {
-	        if (size == 0)
-	            return null;
-	        int s = --size;
-	        ManagedObjectBase result = queue[0];
-	        ManagedObjectBase x = queue[s];
-	        queue[s] = null;
-	        if (s != 0)
-	        	siftDownUsingComparator(0, x);
-	        return result;
-	    }
-
-	    public boolean add(ManagedObjectBase e) {
-	        if (e == null)
-	            throw new NullPointerException();
-	        int i = size;
-	        if (i >= queue.length)
-	            grow(i + 1);
-	        size = i + 1;
-	        if (i == 0)
-	            queue[0] = e;
-	        else
-	        	siftUpUsingComparator(i, e);
-	        return true;
-	    }	    
-
-	    private ManagedObjectBase removeAt(int i) {
-	        // assert i >= 0 && i < size;
-	        int s = --size;
-	        if (s == i) // removed last element
-	            queue[i] = null;
-	        else {
-	        	ManagedObjectBase moved = queue[s];
-	            queue[s] = null;
-	            siftDownUsingComparator(i, moved);
-	            if (queue[i] == moved) {
-	            	siftUpUsingComparator(i, moved);
-	                if (queue[i] != moved)
-	                    return moved;
-	            }
-	        }
-	        return null;
-	    }
-	    
-	    private void siftUpUsingComparator(int k, ManagedObjectBase x) {
-	        while (k > 0) {
-	            int parent = (k - 1) >>> 1;
-	            ManagedObjectBase e = queue[parent];
-	            if (comparator.compare(x, e) >= 0)
-	                break;
-	            queue[k] = e;
-	            e.candleIndex = k;
-	            k = parent;
-	        }
-	        queue[k] = x;
-	        x.candleIndex = k;
-	    }    
-	    
-	    private void siftDownUsingComparator(int k, ManagedObjectBase x) {
-	        int half = size >>> 1;
-	        while (k < half) {
-	            int child = (k << 1) + 1;
-	            ManagedObjectBase c = queue[child];
-	            int right = child + 1;
-	            if (right < size && comparator.compare(c, queue[right]) > 0)
-	                c = queue[child = right];
-	            if (comparator.compare(x, c) <= 0)
-	                break;
-	            queue[k] = c;
-	            c.candleIndex = k;
-	            k = child;
-	        }
-	        queue[k] = x;
-	        x.candleIndex = k;
-	    }
-	    
-	    public int size() 
-	    {
-	    	return this.size;
-	    }
 	}
 }
