+++ java/client/src/main/java/asyncCache/client/BinarySerializerBase.java
@@ -1,4 +1,4 @@
-package asyncMemManager.common;
+package asyncCache.client;
 
 import java.util.function.Function;
