+++ java/common/src/main/java/asyncMemManager/common/FlowKeyConfiguration.java
@@ -1,0 +1,6 @@
+package asyncMemManager.common;
+
+
+public class FlowKeyConfiguration
+{	
+}
