+++ java/common/src/main/java/asyncMemManager/common/di/IndexableQueuedObject.java
@@ -1,0 +1,5 @@
+package asyncMemManager.common.di;
+
+public interface IndexableQueuedObject {
+	public void setIndexInQueue(int idx);
+}
