+++ java/pom.xml
@@ -9,7 +9,7 @@
 	<name>asyncMemManager</name>
 	<description>Async Mem Manager</description>
 	<properties>
-    <asyncCache.version>0.0.1-SNAPSHOT</asyncCache.version>
+    <asyncMemManager.version>0.0.1-SNAPSHOT</asyncMemManager.version>
 		<maven.compiler.target>1.8</maven.compiler.target>
 		<maven.compiler.source>1.8</maven.compiler.source>
     <java.version>1.8</java.version>
